python/requirements.txt
prometheus_client==0.17.1
psycopg2-binary==2.9.7
requests==2.31.0
smtplib2==0.2.1